moved lib directory within /src directory
