+++ client/src/lib/supabase/auth/handleGoogleLoginClientSide.ts
