+++ client/src/lib/supabase/auth/handleGoogleLoginServerSide.ts
