+++ client/src/lib/supabase/middleware.ts
